+++ Editor/UI/Framework/AnalysisView.cs
@@ -31,6 +31,9 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
         protected ViewManager m_ViewManager;
 
         DependencyView m_DependencyView;
+        bool m_ShowInfo;
+        bool m_ShowWarn;
+        bool m_ShowError;
         GUIContent m_HelpButtonContent;
         IssueTable m_Table;
         IssueLayout m_Layout;
@@ -55,6 +58,8 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
 
         public AnalysisView(ViewManager viewManager)
         {
+            m_ShowInfo = m_ShowWarn = m_ShowError = true;
+
             m_ViewManager = viewManager;
         }
 
@@ -310,6 +315,18 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
                     Refresh();
                 }
             }
+
+            if (m_Desc.showSeverityFilters)
+            {
+                EditorGUI.BeginChangeCheck();
+                m_ShowInfo = GUILayout.Toggle(m_ShowInfo, Utility.InfoIcon, EditorStyles.toolbarButton, GUILayout.ExpandWidth(false));
+                m_ShowWarn = GUILayout.Toggle(m_ShowWarn, Utility.WarnIcon, EditorStyles.toolbarButton, GUILayout.ExpandWidth(false));
+                m_ShowError = GUILayout.Toggle(m_ShowError, Utility.ErrorIcon, EditorStyles.toolbarButton, GUILayout.ExpandWidth(false));
+                if (EditorGUI.EndChangeCheck())
+                {
+                    Refresh();
+                }
+            }
         }
 
         void DrawDataOptions()
@@ -439,6 +456,24 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
 
         public virtual bool Match(ProjectIssue issue)
         {
+            if (m_Desc.showSeverityFilters)
+            {
+                switch (issue.descriptor.severity)
+                {
+                    case Rule.Severity.Info:
+                        if (!m_ShowInfo)
+                            return false;
+                        break;
+                    case Rule.Severity.Warning:
+                        if (!m_ShowWarn)
+                            return false;
+                        break;
+                    case Rule.Severity.Error:
+                        if (!m_ShowError)
+                            return false;
+                        break;
+                }
+            }
             return m_BaseFilter.Match(issue) && m_TextFilter.Match(issue);
         }