+++ Tests/Editor/CodeAnalysisTests.cs
@@ -25,7 +25,8 @@ namespace Unity.ProjectAuditor.EditorTests
         TempAsset m_TempAssetIssueInVirtualMethod;
         TempAsset m_TempAssetAnyApiInNamespace;
         TempAsset m_TempAssetObjectName;
-        TempAsset m_TempAssetRendererMaterial;
+        TempAsset m_TempAssetGenericInstantiation;
+		TempAsset m_TempAssetRendererMaterial;
 
         [OneTimeSetUp]
         public void SetUp()
@@ -209,6 +210,18 @@ class ObjectNameTest : MonoBehaviour
         Debug.Log(gameObject.name);
         Debug.Log(transform.name);
         Debug.Log(this.name);
+    }
+}
+");
+
+            m_TempAssetGenericInstantiation = new TempAsset("GenericInstantiation.cs", @"
+using System.Collections.Generic;
+class GenericInstantiation
+{
+    HashSet<string> m_GenericInstance;
+    void Dummy()
+    {
+        m_GenericInstance = new HashSet<string>();
     }
 }
 ");
@@ -432,6 +445,15 @@ class RendererMaterialTest
             Assert.True(issues.All(i => i.description.Equals("UnityEngine.Object.name")));
         }
 
+        [Test]
+        public void CodeAnalysis_GenericInstantiation_IsReported()
+        {
+            var issues = Utility.AnalyzeAndFindAssetIssues(m_TempAssetGenericInstantiation, IssueCategory.GenericInstance);
+
+            Assert.AreEqual(1, issues.Length);
+            Assert.AreEqual("System.Collections.Generic.HashSet`1<System.String>", issues.First().description);
+        }
+		
         [Test]
         public void CodeAnalysis_RendererMaterial_IsReported()
         {