+++ Tests/Editor/ShadersAnalysisTests.cs
@@ -419,8 +419,10 @@ Shader ""Custom/MyEditorShader""
             Assert.Positive(variants.Length, "No shader variants found");
 
             // check custom properties
-            Assert.True(variants.Any(v => v.GetCustomProperty(ShaderVariantProperty.Keywords).Equals("<no keywords>")), "No shader variants found without keywords");
-            Assert.True(variants.Any(v => v.GetCustomProperty(ShaderVariantProperty.Requirements).Equals("BaseShaders Derivatives")), "No shader variants found with BaseShaders and Derivatives requirements");
+            Assert.True(variants.Any(v => v.GetCustomProperty(ShaderVariantProperty.Keywords).Equals(ShadersModule.k_NoKeywords)), "No shader variants found without keywords");
+
+            var expectedRequirements = Formatting.CombineStrings(new[] {ShaderRequirements.BaseShaders, ShaderRequirements.Derivatives}.Select(r => r.ToString()).ToArray());
+            Assert.True(variants.Any(v => v.GetCustomProperty(ShaderVariantProperty.Requirements).Equals(expectedRequirements)), "No shader variants found with {0} requirements", expectedRequirements);
 #if UNITY_2019_1_OR_NEWER
             //this one fails on yamato
             //Assert.True(variants.Any(v => v.GetCustomProperty(ShaderVariantProperty.Keywords).Equals("INSTANCING_ON")), "No shader variants found with INSTANCING_ON keyword");